chore: change recommend component modal
