--- src/modules/booking/components/onlineVisitRecommend/index.ts
@@ -1,1 +1,0 @@
-export { default } from './onlineVisitRecommend';
