+++ src/modules/booking/views/index.tsx
@@ -10,7 +10,7 @@ import { useGetProfileData } from '@/common/apis/services/profile/getFullProfile
 // Hooks
 import { useFeatureValue } from '@growthbook/growthbook-react';
 import { useRouter } from 'next/router';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useMemo, useRef, useState } from 'react';
 import { uniqMessengers } from '../functions/uniqMessengers';
 
 // Components
@@ -52,7 +52,9 @@ import { UserInfo } from '@/modules/login/store/userInfo';
 import { useGetNationalCodeConfirmation } from '@/common/apis/services/booking/getNationalCodeConfirmation';
 import { useInquiryIdentityInformation } from '@/common/apis/services/booking/inquiryIdentityInformation';
 import { useUnsuspend } from '@/common/apis/services/booking/unsuspend';
+import { useSearch } from '@/common/apis/services/search/search';
 import Alert from '@/common/components/atom/alert';
+import Loading from '@/common/components/atom/loading';
 import { FakeData } from '@/common/constants/fakeData';
 import useApplication from '@/common/hooks/useApplication';
 import useCustomize from '@/common/hooks/useCustomize';
@@ -60,10 +62,11 @@ import useModal from '@/common/hooks/useModal';
 import { splunkBookingInstance, splunkSearchInstance } from '@/common/services/splunk';
 import classNames from '@/common/utils/classNames';
 import { convertNumberToStringGender } from '@/common/utils/convertNumberToStringGender';
+import SearchCard from '@/modules/search/components/card/card';
 import { useSearchRouting } from '@/modules/search/hooks/useSearchRouting';
 import axios from 'axios';
 import moment from 'jalali-moment';
-import OnlineVisitRecommend from '../components/onlineVisitRecommend';
+import random from 'lodash/random';
 import { defaultMessengers } from '../constants/defaultMessengers';
 import { reformattedCentersProperty } from '../functions/reformattedCentersProperty';
 import { reformattedServicesProperty } from '../functions/reformattedServicesProperty';
@@ -136,7 +139,6 @@ const BookingSteps = (props: BookingStepsProps) => {
   const shouldUseInquiryIdentityInformation = useFeatureValue<{ ids: string[] }>('booking:inquiry-identity-information|center-list', {
     ids: [],
   });
-
   const onlineVisitDoctorList = useFeatureValue<{ slugs: string[] }>('booking:online-visit-recommend-modal', {
     slugs: [],
   });
@@ -174,7 +176,13 @@ const BookingSteps = (props: BookingStepsProps) => {
     serviceId: service?.id,
     userCenterId: center?.user_center_id,
   });
-
+  const searchData = useSearch({
+    route: decodeURIComponent(`ir/${profile.expertises[0]?.expertise_groups[0].en_slug}`),
+    query: {
+      turn_type: 'consult',
+    },
+  });
+  const substituteDoctor = useMemo(() => searchData.data?.search?.result?.[random(0, 2)] ?? {}, [searchData.data]);
   const [step, setStep] = useState<Step>(defaultStep?.step ?? 'SELECT_CENTER');
 
   useEffect(() => {
@@ -425,6 +433,21 @@ const BookingSteps = (props: BookingStepsProps) => {
       buttons,
     });
     handleOpenErrorModal();
+  };
+
+  const handleClickDcotorCardDoctor = ({ url }: { url: string }) => {
+    splunkSearchInstance().sendEvent({
+      group: 'booking-freeturn-error',
+      type: 'booking-freeturn-error-click-doctor-card',
+      event: {
+        slug: url.replace('/dr/', ''),
+        doctor_name: profile?.display_name,
+        doctor_expertice: profile?.expertises[0]?.expertise_groups[0]?.name,
+        center_id: center?.id,
+        service_id: service?.id,
+      },
+    });
+    location.assign(url.replace('/dr/', '/booking/') + '?centerId=5532&skipTimeSelectStep=true');
   };
 
   return (
@@ -822,32 +845,48 @@ const BookingSteps = (props: BookingStepsProps) => {
                       <Alert severity="success" className="p-3 text-green-700 text-sm font-medium">
                         بدون خروج از منزل، آنلاین ویزیت شوید.
                       </Alert>
-                      <OnlineVisitRecommend
-                        route={`ir/${profile.expertises[0]?.expertise_groups[0].en_slug}`}
-                        turn_type="consult"
-                        classnames="shadow-none !py-4 lg:!py-4 cursor-pointer"
-                        avatarSize="lg"
-                        type="doctor"
-                        details={{
-                          badges: [
-                            {
-                              title: 'تضمین بازپرداخت مبلغ ویزیت در صورت نارضایتی',
-                              icon: 'shield-icon',
-                              type: 'error',
-                            },
-                          ],
-                        }}
-                        event={{
-                          group: 'booking-freeturn-error',
-                          type: 'booking-freeturn-error-click-doctor-card',
-                          data: {
-                            doctor_name: profile?.display_name,
-                            doctor_expertice: profile?.expertises[0]?.expertise_groups[0]?.name,
-                            center_id: center?.id,
-                            service_id: service?.id,
-                          },
-                        }}
-                      />
+                      <div onClick={() => handleClickDcotorCardDoctor({ url: substituteDoctor.url })}>
+                        {(searchData.isLoading || !substituteDoctor?.url) && (
+                          <div className="flex justify-center w-full">
+                            <Loading className="w-8 h-8 my-8 " />
+                          </div>
+                        )}
+                        {searchData.isSuccess && substituteDoctor?.url && (
+                          <SearchCard
+                            avatarSize="lg"
+                            baseInfo={{
+                              displayName: substituteDoctor.title,
+                              expertise: substituteDoctor.display_expertise,
+                              experience: substituteDoctor.experience,
+                              isVerify: true,
+                              avatar: substituteDoctor.image,
+                              rate: {
+                                count: substituteDoctor.rates_count,
+                                satisfaction: substituteDoctor.satisfaction,
+                              },
+                              isOnline: true,
+                            }}
+                            details={{
+                              badges: [
+                                {
+                                  title: 'تضمین بازپرداخت مبلغ ویزیت در صورت نارضایتی',
+                                  icon: 'shield-icon',
+                                  type: 'error',
+                                },
+                              ],
+                            }}
+                            className="shadow-none !py-2 lg:!py-2 cursor-pointer"
+                            type="doctor"
+                            actions={[
+                              {
+                                text: `گفتگو با ${substituteDoctor.title}`,
+                                outline: false,
+                                description: '',
+                              },
+                            ]}
+                          />
+                        )}
+                      </div>
                       <Button block size="sm" className="text-xs opacity-70" variant="text" onClick={handleClickMoreDoctors}>
                         مشاهده سایر پزشکان آنلاین {profile.expertises[0]?.expertise_groups[0].name}
                       </Button>
